TRAANode: New TRAA implementation.
