+++ examples/webgpu_postprocessing_traa.html
@@ -25,8 +25,8 @@
 		<script type="module">
 
 			import * as THREE from 'three';
-			import { mrt, output, velocity } from 'three/tsl';
-			import { traaPass } from 'three/addons/tsl/display/TRAAPassNode.js';
+			import { mrt, output, pass, velocity } from 'three/tsl';
+			import { traa } from 'three/addons/tsl/display/TRAANode.js';
 
 			import Stats from 'three/addons/libs/stats.module.js';
 
@@ -74,13 +74,20 @@
 				// postprocessing
 
 				postProcessing = new THREE.PostProcessing( renderer );
-				const scenePass = traaPass( scene, camera );
+				const scenePass = pass( scene, camera );
 				scenePass.setMRT( mrt( {
 					output: output,
 					velocity: velocity
 				} ) );
 
-				postProcessing.outputNode = scenePass;
+				const scenePassColor = scenePass.getTextureNode( 'output' );
+				const scenePassDepth = scenePass.getTextureNode( 'depth' );
+				const scenePassVelocity = scenePass.getTextureNode( 'velocity' );
+
+				const traaNode = traa( scenePassColor, scenePassDepth, scenePassVelocity, camera );
+			
+				postProcessing.outputNode = traaNode;
+				postProcessing.setTRAANode( traaNode );
 
 				//
