+++ test/e2e/puppeteer.js
@@ -172,6 +172,7 @@ const exceptionList = [
 	'webgpu_textures_2d-array_compressed',
 	'webgpu_rendertarget_2d-array_3d',
 	'webgpu_materials_envmaps_bpcem',
+	'webgpu_postprocessing_ao',
 	'webgpu_postprocessing_sobel',
 	'webgpu_postprocessing_3dlut',
 	'webgpu_postprocessing_fxaa',
